[fix] ログファイルを ZIP で出力する際、既存のファイルを削除できない不具合を修正

+++ MyPad/ViewModels/Regions/OptionContentViewModel.cs
@@ -101,6 +101,9 @@ namespace MyPad.ViewModels.Regions
         [LogInterceptor]
         private async Task<bool> ExportLogArchive(string path)
         {
+            const int MAX_LOOP_COUNT = 10;
+            const int LOOP_DELAY = 500;
+
             var sourcePath = ((App)Application.Current).LogDirectoryPath;
             var tempPath = Path.Combine(this.ProductInfo.Temporary, Path.GetFileNameWithoutExtension(path));
 
@@ -108,18 +111,30 @@ namespace MyPad.ViewModels.Regions
             {
                 this.IsWorking.Value = true;
 
-                await Task.Run(() =>
-                {
-                    FileSystem.CopyDirectory(sourcePath, tempPath, UIOption.AllDialogs, UICancelOption.ThrowException);
-                    ZipFile.CreateFromDirectory(tempPath, path, CompressionLevel.Optimal, false);
-                });
-                _ = Task.Run(() => Directory.Delete(tempPath, true));
+                // 一時フォルダに退避する
+                if (Directory.Exists(tempPath))
+                    await Task.Run(() => Directory.Delete(tempPath, true));
+                for (var i = 0; i < MAX_LOOP_COUNT && Directory.Exists(tempPath); i++)
+                    await Task.Delay(LOOP_DELAY);
+                if (Directory.Exists(tempPath))
+                    throw new IOException(Resources.Message_NotifyTryAgainLater);
+                await Task.Run(() => FileSystem.CopyDirectory(sourcePath, tempPath, UIOption.AllDialogs, UICancelOption.ThrowException));
+
+                // 退避したファイルを圧縮して出力する
+                if (File.Exists(path))
+                    await Task.Run(() => File.Delete(path));
+                for (var i = 0; i < MAX_LOOP_COUNT && File.Exists(path); i++)
+                    await Task.Delay(LOOP_DELAY);
+                if (File.Exists(path))
+                    throw new IOException(Resources.Message_NotifyTryAgainLater);
+                await Task.Run(() => ZipFile.CreateFromDirectory(tempPath, path, CompressionLevel.Optimal, false));
 
                 Process.Start("explorer.exe", $"/select, {path}");
                 this.Logger.Log($"ログファイルを出力しました。: Path={path}, Temp={tempPath}", Category.Info);
             }
             catch (OperationCanceledException e)
             {
+                // MEMO: FileSystem.CopyDirectory の処理をキャンセルした場合
                 this.DialogService.Notify(e.Message);
                 return false;
             }
@@ -131,6 +146,8 @@ namespace MyPad.ViewModels.Regions
             }
             finally
             {
+                if (Directory.Exists(tempPath))
+                    _ = Task.Run(() => { try { Directory.Delete(tempPath, true); } catch { } });
                 this.IsWorking.Value = false;
             }
             return true;
